# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      id_departamento:
        type: integer
      nome:
        type: string
      senha:
        type: string
      tipo_conta:
        type: string
    type: object
info:
  contact: {}
  description: API do sistema de ressarcimento (SURE)
  title: SURE API
  version: "1.0"
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Auth
...
  /api/v1/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar usuario
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"